🦄 refactor(examples):
move related codes from client to model
