+++ examples/federated_learning/client.py
@@ -3,9 +3,7 @@ import numpy as np
 import os
 import mnist
 import pytorch_lightning as pl
-from collections import OrderedDict
 
-import torch
 from torch.utils.data import DataLoader, random_split, Subset
 from torchvision import transforms
 from torchvision.datasets import MNIST
@@ -37,30 +35,6 @@ def load_data(id):
 
     return train_loader, val_loader, test_loader
 
-def load_parameters(parameters_path):
-    try:
-        parameters = np.load(parameters_path, allow_pickle=True).tolist()
-        return parameters
-    except OSError:
-        return None
-
-def get_parameters(model):
-    def _get_parameters(m):
-        return [val.cpu().numpy() for _, val in m.state_dict().items()]
-    encoder_params = _get_parameters(model.encoder)
-    decoder_params = _get_parameters(model.decoder)
-    return encoder_params + decoder_params
-
-def set_parameters(model, parameters):
-    def _set_parameters(m, p):
-        params_dict = zip(m.state_dict().keys(), p)
-        state_dict = OrderedDict({k: torch.tensor(v) for k, v in params_dict})
-        m.load_state_dict(state_dict, strict=True)
-        
-    if parameters != None:
-        _set_parameters(model.encoder, parameters[:4])
-        _set_parameters(model.decoder, parameters[4:])
-
 def main() -> None:
     parser = argparse.ArgumentParser(description='PyTorch MNIST Example -- Client')
     parser.add_argument('--clientID', type=int, default=0, metavar='N',
@@ -79,8 +53,8 @@ def main() -> None:
     train_loader, val_loader, test_loader = load_data(args.clientID)
 
     # initial parameters
-    parameters = load_parameters(args.path)
-    set_parameters(model, parameters)
+    parameters = mnist.load_parameters(args.path)
+    mnist.set_parameters(model, parameters)
 
     # train
     trainer = pl.Trainer(max_epochs=1, max_steps=10, enable_progress_bar=True)
@@ -93,8 +67,8 @@ def main() -> None:
         np.savetxt(f, [loss])
 
     # save
-    parameters = get_parameters(model)
-    np.save(dir+'parameters.npy', np.array(parameters, dtype=object))
+    parameters = mnist.get_parameters(model)
+    mnist.save_parameters(dir+'parameters.npy')
   
 if __name__ == "__main__":
     main()
